feat: confirmation on recipe delete

+++ app.js
@@ -55,9 +55,26 @@ form.addEventListener("submit", (e) => {
 recipeList.addEventListener("click", (e) => {
     const target = e.target;
     if (target.tagName === "BUTTON") {
-        const recipeID = target.parentElement.getAttribute("data-id");
-        db.collection("recipes").doc(recipeID).delete().then(() => {
-            console.log("recipe deleted");
-        }).catch((err) => console.log(err));
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                const recipeID = target.parentElement.getAttribute("data-id");
+                db.collection("recipes").doc(recipeID).delete().then(() => {
+                    console.log("recipe deleted");
+                }).catch((err) => console.log(err));
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+            }
+        });
     }
 })
